cn0326: made to calibrate before every data read and fixed two bugs
To work properly, the sensor has to be calibrated before every data
aquisition. The first bug was the wait condition in the scan method where
the program was waiting for convertion to finish only in continuous
convertion mode. The second bus was that the wait condition for the
convertion was wrong. These changes address these issues.

Signed-off-by: adrimbarean <Andrei.Drimbarean@analog.com>